requirements.txt
python-telegram-bot==12.4.2
emoji
git+https://gitlab.com/HirschHeissIch/ptbstats.git@v1.1
